refactor: move the activity layout code to the fragment
Change from using the activity for the main screen of the application, to instead use the
PusherFragment
