+++ feature/pusher/build.gradle
@@ -19,6 +19,9 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+    buildFeatures {
+        viewBinding true
+    }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
@@ -30,16 +33,19 @@ android {
 
 dependencies {
 
+    implementation project(":library:common")
     implementation project(":library:datasource:pwpush")
     implementation project(":library:resources")
 
     implementation "androidx.appcompat:appcompat:1.3.0"
     implementation "androidx.core:core-ktx:1.5.0"
+    implementation "androidx.fragment:fragment-ktx:1.3.4"
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.3.1"
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.3.1"
     implementation "com.chesire.lintrules:lint-gradle:1.2.6"
     implementation "com.chesire.lintrules:lint-xml:1.2.6"
     implementation "com.google.android.material:material:1.3.0"
+    implementation "com.squareup.okhttp3:okhttp:4.9.1"
 
     testImplementation "junit:junit:4.13.2"
 }
